+++ google-cloud-spanner/src/main/java/com/google/cloud/spanner/Instance.java
@@ -23,6 +23,7 @@ import com.google.cloud.Timestamp;
 import com.google.cloud.spanner.Options.ListOption;
 import com.google.longrunning.Operation;
 import com.google.spanner.admin.database.v1.CreateDatabaseMetadata;
+import com.google.spanner.admin.instance.v1.AutoscalingConfig;
 import com.google.spanner.admin.instance.v1.UpdateInstanceMetadata;
 import java.util.Map;
 
@@ -83,6 +84,12 @@ public class Instance extends InstanceInfo {
     @Override
     public Builder setProcessingUnits(int processingUnits) {
       infoBuilder.setProcessingUnits(processingUnits);
+      return this;
+    }
+
+    @Override
+    public Builder setAutoscalingConfig(AutoscalingConfig autoscalingConfig) {
+      infoBuilder.setAutoscalingConfig(autoscalingConfig);
       return this;
     }
 
@@ -220,6 +227,7 @@ public class Instance extends InstanceInfo {
             .setNodeCount(proto.getNodeCount())
             .setCreateTime(Timestamp.fromProto(proto.getCreateTime()))
             .setUpdateTime(Timestamp.fromProto(proto.getUpdateTime()))
+            .setAutoscalingConfig(proto.getAutoscalingConfig())
             .setProcessingUnits(proto.getProcessingUnits());
     State state;
     switch (proto.getState()) {